requirements.txt
# Web Framework
flask==2.3.3
gunicorn==21.2.0

# Message Queue
pika==1.3.2
celery==5.3.4

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Cache
redis==5.0.1

# Google Drive API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# AWS
boto3==1.34.0
botocore==1.34.0

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Scheduling
schedule==1.2.0

# Data Handling
dataclasses-json==0.6.3

# Monitoring
prometheus-client==0.19.0
prometheus-flask-exporter==0.23.0

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3